Added Zuul headers
